url-shortener-api/requirements.txt
#All latest versions
pymongo
pymongo[srv]
python-dotenv
Flask

#All dependencies specified in corresponding docker file